registro-notas/pubsub-service/requirements.txt
google-cloud-pubsub
tornado
